Fix Tests on 32bit Platform
- Change Int64->Int for general readtable() case
- Force rand to generate Int64 values due to an inconsistency in selecting
  random elements

+++ test/data.jl
@@ -94,8 +94,9 @@ module TestData
     #test_group("DataFrame")
     srand(1)
     N = 20
-    d1 = pdata(rand(1:2, N))
-    d2 = (@pdata ["A", "B", NA])[rand(1:3, N)]
+    #Cast to int64 as rand() behavior differs between Int32/64
+    d1 = pdata(rand(int64(1:2), N))
+    d2 = (@pdata ["A", "B", NA])[rand(int64(1:3), N)]
     d3 = data(randn(N))
     d4 = data(randn(N))
     df7 = DataFrame({d1, d2, d3}, [:d1, :d2, :d3])
